+++ src/Command/Mount/MountDownloadCommand.php
@@ -38,10 +38,7 @@ class MountDownloadCommand extends MountCommandBase
         $this->validateInput($input);
 
         $appName = $this->selectApp($input);
-        $sshUrl = $this->getSelectedEnvironment()
-            ->getSshUrl($appName);
-
-        $appConfig = $this->getAppConfig($sshUrl, (bool) $input->getOption('refresh'));
+        $appConfig = $this->getAppConfig($appName, (bool) $input->getOption('refresh'));
 
         if (empty($appConfig['mounts'])) {
             $this->stdErr->writeln(sprintf('The app "%s" doesn\'t define any mounts.', $appConfig['name']));
@@ -123,7 +120,7 @@ class MountDownloadCommand extends MountCommandBase
             return 1;
         }
 
-        $this->runSync($sshUrl, $mountPath, $target, false, [
+        $this->runSync($appName, $mountPath, $target, false, [
             'delete' => $input->getOption('delete'),
             'exclude' => $input->getOption('exclude'),
             'include' => $input->getOption('include'),