+++ src/Service/Api.php
@@ -196,8 +196,7 @@ class Api
             $session->setId('cli-' . $this->sessionId);
 
             $this->sessionStorage = KeychainStorage::isSupported()
-                && $this->config->has('experimental.use_keychain')
-                && $this->config->get('experimental.use_keychain')
+                && $this->config->isExperimentEnabled('use_keychain')
                 ? new KeychainStorage($this->config->get('application.name'))
                 : new File($this->config->getWritableUserDir() . '/.session');
             $session->setStorage($this->sessionStorage);